Add master control, start networking code
Add the master control thread that all other tasks will notify (and will
control the behaviour of the device).

Currently, after 1 minute of scanning on BLE, we'll initiate a
connection to the backend and deliver a polite greeting. Next step is to
wire up delivering the protobufs containing the BLE device data, free up
the associated resources.

The inititial threads of the failure state management are present as
well, but need to be filled in. When we can't connect to the AP, fail to
connect to the backend or are high memory pressure, we'll force a loop
of attempts to find the AP, while sleeping for 1 minute each time. We'll
also set the failure LED high, so that anyone looking at the device will
know it is having connectivity errors.